cleanup invite test (and make it fail :(
